move a fixture out to conftest.py
